Refactor transparency log
